add drawable component to wrap the various sprite types
