+++ assets/decoration/seaweed.json
@@ -2,6 +2,7 @@
   "id": "seaweed",
   "sprite": {
     "texture": "default_decoration",
+    "autoplay_id": "idle",
     "frame_size": {
       "x": 48,
       "y": 51
