+++ assets/items/cannon.json
@@ -51,6 +51,7 @@
       ],
       "sprite": {
         "texture": "cannon_ball",
+        "autoplay_id": "effect",
         "animations": [
           {
             "id": "effect",
@@ -59,13 +60,13 @@
             "fps": 1,
             "is_looping": true
           }
-        ],
-        "should_autoplay": true
+        ]
       }
     }
   ],
   "sprite": {
     "texture": "cannon",
+    "autoplay_id": "idle",
     "animations": [
       {
         "id": "idle",
