+++ assets/items/mines.json
@@ -42,6 +42,7 @@
           "timed_trigger": 0.7,
           "sprite": {
             "texture": "explosion_small",
+            "autoplay_id": "effect",
             "animations": [
               {
                 "id": "effect",
@@ -56,6 +57,7 @@
       "collide_with_platforms": true,
       "sprite": {
         "texture": "mines",
+        "autoplay_id": "effect",
         "animations": [
           {
             "id": "effect",
@@ -64,8 +66,7 @@
             "fps": 12,
             "is_looping": true
           }
-        ],
-        "should_autoplay": true
+        ]
       }
     }
   ],
@@ -83,6 +84,7 @@
   },
   "sprite": {
     "texture": "mines",
+    "autoplay_id": "idle",
     "animations": [
       {
         "id": "idle",
