+++ assets/items/sniper_rifle.json
@@ -55,6 +55,7 @@
   ],
   "sprite": {
     "texture": "sniper_rifle",
+    "autoplay_id": "idle",
     "animations": [
       {
         "id": "idle",
@@ -67,7 +68,7 @@
         "id": "attack",
         "row": 1,
         "frames": 3,
-        "fps": 15
+        "fps": 14
       }
     ]
   },
