+++ src/map/sproinger.rs
@@ -1,11 +1,12 @@
 use macroquad::audio::play_sound_once;
 use macroquad::experimental::collections::storage;
 use macroquad::prelude::*;
+use std::borrow::BorrowMut;
 
 use hecs::{Entity, World};
 
 use crate::{
-    AnimatedSprite, Animation, DrawOrder, PhysicsBody, QueuedAnimationAction, Resources, Result,
+    Animation, Drawable, DrawableKind, PhysicsBody, QueuedAnimationAction, Resources, Result,
     Transform,
 };
 
@@ -40,13 +41,6 @@ impl Sproinger {
 }
 
 pub fn spawn_sproinger(world: &mut World, position: Vec2) -> Result<Entity> {
-    let texture_res = {
-        let resources = storage::get::<Resources>();
-        resources.textures.get(TEXTURE_ID).cloned().unwrap()
-    };
-
-    let frame_size = texture_res.frame_size();
-
     let animations = &[
         Animation {
             id: IDLE_ANIMATION_ID.to_string(),
@@ -74,10 +68,9 @@ pub fn spawn_sproinger(world: &mut World, position: Vec2) -> Result<Entity> {
     let entity = world.spawn((
         Sproinger::new(),
         Transform::from(position),
-        DrawOrder(SPROINGER_DRAW_ORDER),
-        AnimatedSprite::new(
-            texture_res.texture,
-            frame_size,
+        Drawable::new_animated_sprite(
+            SPROINGER_DRAW_ORDER,
+            TEXTURE_ID,
             animations,
             Default::default(),
         ),
@@ -103,8 +96,8 @@ pub fn update_sproingers(world: &mut World) {
 
     let mut to_be_sproinged = Vec::new();
 
-    'sproingers: for (_, (sproinger, transform, sprite)) in
-        world.query_mut::<(&mut Sproinger, &Transform, &mut AnimatedSprite)>()
+    'sproingers: for (_, (sproinger, transform, drawable)) in
+        world.query_mut::<(&mut Sproinger, &Transform, &mut Drawable)>()
     {
         sproinger.cooldown_timer += dt;
 
@@ -114,26 +107,28 @@ pub fn update_sproingers(world: &mut World) {
         };
 
         if sproinger.cooldown_timer >= COOLDOWN {
-            sprite.set_animation(IDLE_ANIMATION_ID, true);
+            if let DrawableKind::AnimatedSprite(sprite) = drawable.kind.borrow_mut() {
+                sprite.set_animation(IDLE_ANIMATION_ID, true);
 
-            let position = transform.position - (sprite.frame_size / 2.0);
+                let position = transform.position - (sprite.frame_size / 2.0);
 
-            let trigger_rect = Rect::new(position.x, position.y, TRIGGER_WIDTH, TRIGGER_HEIGHT);
+                let trigger_rect = Rect::new(position.x, position.y, TRIGGER_WIDTH, TRIGGER_HEIGHT);
 
-            for (e, rect) in &bodies {
-                if trigger_rect.overlaps(rect) {
-                    to_be_sproinged.push(*e);
+                for (e, rect) in &bodies {
+                    if trigger_rect.overlaps(rect) {
+                        to_be_sproinged.push(*e);
 
-                    sproinger.cooldown_timer = 0.0;
+                        sproinger.cooldown_timer = 0.0;
 
-                    sprite.set_animation(EXPAND_ANIMATION_ID, true);
-                    sprite.queue_action(QueuedAnimationAction::Play(
-                        CONTRACT_ANIMATION_ID.to_string(),
-                    ));
+                        sprite.set_animation(EXPAND_ANIMATION_ID, true);
+                        sprite.queue_action(QueuedAnimationAction::Play(
+                            CONTRACT_ANIMATION_ID.to_string(),
+                        ));
 
-                    play_sound_once(sound);
+                        play_sound_once(sound);
 
-                    continue 'sproingers;
+                        continue 'sproingers;
+                    }
                 }
             }
         }
