+++ src/player/inventory.rs
@@ -1,4 +1,5 @@
 use macroquad::prelude::*;
+use std::borrow::BorrowMut;
 
 use hecs::{Entity, With, Without, World};
 
@@ -8,9 +9,7 @@ use crate::items::{
 };
 use crate::particles::ParticleEmitter;
 use crate::player::{PlayerController, PlayerState, IDLE_ANIMATION_ID, PICKUP_GRACE_TIME};
-use crate::{
-    AnimatedSpriteSet, DrawOrder, Item, Owner, PassiveEffectInstance, PhysicsBody, Transform,
-};
+use crate::{Drawable, DrawableKind, Item, Owner, PassiveEffectInstance, PhysicsBody, Transform};
 
 const THROW_FORCE: f32 = 5.0;
 
@@ -90,8 +89,11 @@ pub fn update_player_inventory(world: &mut World) {
                     let mut body = world.get_mut::<PhysicsBody>(item_entity).unwrap();
                     body.is_deactivated = true;
 
-                    let mut sprite_set = world.get_mut::<AnimatedSpriteSet>(item_entity).unwrap();
-                    sprite_set.set_all(IDLE_ANIMATION_ID, true);
+                    let mut drawable = world.get_mut::<Drawable>(item_entity).unwrap();
+                    if let DrawableKind::AnimatedSpriteSet(sprite_set) = drawable.kind.borrow_mut()
+                    {
+                        sprite_set.set_all(IDLE_ANIMATION_ID, true);
+                    }
 
                     continue;
                 }
@@ -124,9 +126,12 @@ pub fn update_player_inventory(world: &mut World) {
                             let mut body = world.get_mut::<PhysicsBody>(weapon_entity).unwrap();
                             body.is_deactivated = true;
 
-                            let mut sprite_set =
-                                world.get_mut::<AnimatedSpriteSet>(weapon_entity).unwrap();
-                            sprite_set.set_all(IDLE_ANIMATION_ID, true);
+                            let mut drawable = world.get_mut::<Drawable>(weapon_entity).unwrap();
+                            if let DrawableKind::AnimatedSpriteSet(sprite_set) =
+                                drawable.kind.borrow_mut()
+                            {
+                                sprite_set.set_all(IDLE_ANIMATION_ID, true);
+                            }
 
                             break;
                         }
@@ -146,16 +151,19 @@ pub fn update_player_inventory(world: &mut World) {
 
                 weapon_transform.position = weapon_mount;
 
-                let mut sprite_set = world.get_mut::<AnimatedSpriteSet>(weapon_entity).unwrap();
+                let mut drawable = world.get_mut::<Drawable>(weapon_entity).unwrap();
+                let mut frame_size = Vec2::ZERO;
 
-                sprite_set.flip_all_x(state.is_facing_left);
-                sprite_set.flip_all_y(state.is_upside_down);
+                if let DrawableKind::AnimatedSpriteSet(sprite_set) = drawable.kind.borrow_mut() {
+                    sprite_set.flip_all_x(state.is_facing_left);
+                    sprite_set.flip_all_y(state.is_upside_down);
 
-                let frame_size = sprite_set
-                    .map
-                    .get(SPRITE_ANIMATED_SPRITE_ID)
-                    .map(|sprite| sprite.size())
-                    .unwrap();
+                    frame_size = sprite_set
+                        .map
+                        .get(SPRITE_ANIMATED_SPRITE_ID)
+                        .map(|sprite| sprite.size())
+                        .unwrap();
+                }
 
                 let mount_offset = flip_offset(
                     weapon.mount_offset,
@@ -239,21 +247,25 @@ pub fn update_player_inventory(world: &mut World) {
                 } else {
                     let position = transform.position;
 
-                    match world.get_mut::<AnimatedSpriteSet>(item_entity) {
-                        Ok(mut sprite_set) => {
-                            sprite_set.flip_all_x(state.is_facing_left);
-                            sprite_set.flip_all_y(state.is_upside_down);
+                    match world.get_mut::<Drawable>(item_entity) {
+                        Ok(mut drawable) => {
+                            if let DrawableKind::AnimatedSpriteSet(sprite_set) =
+                                drawable.kind.borrow_mut()
+                            {
+                                sprite_set.flip_all_x(state.is_facing_left);
+                                sprite_set.flip_all_y(state.is_upside_down);
 
-                            let offset = flip_offset(
-                                item.mount_offset,
-                                sprite_set.size(),
-                                state.is_facing_left,
-                                state.is_upside_down,
-                            );
+                                let offset = flip_offset(
+                                    item.mount_offset,
+                                    sprite_set.size(),
+                                    state.is_facing_left,
+                                    state.is_upside_down,
+                                );
 
-                            let mut item_transform =
-                                world.get_mut::<Transform>(item_entity).unwrap();
-                            item_transform.position = position + offset;
+                                let mut item_transform =
+                                    world.get_mut::<Transform>(item_entity).unwrap();
+                                item_transform.position = position + offset;
+                            }
                         }
                         Err(err) => {
                             #[cfg(debug_assertions)]
@@ -270,10 +282,13 @@ pub fn update_player_inventory(world: &mut World) {
     for (player_entity, item_entity) in picked_up {
         world.insert_one(item_entity, Owner(player_entity)).unwrap();
 
-        let player_draw_order = world.get::<DrawOrder>(player_entity).unwrap();
+        let player_draw_order = world
+            .get::<Drawable>(player_entity)
+            .map(|drawable| drawable.draw_order)
+            .unwrap();
 
-        let mut draw_order = world.get_mut::<DrawOrder>(item_entity).unwrap();
-        draw_order.0 = player_draw_order.0 + 1;
+        let mut drawable = world.get_mut::<Drawable>(item_entity).unwrap();
+        drawable.draw_order = player_draw_order + 1;
 
         if let Ok(item) = world.get::<Item>(item_entity) {
             let mut state = world.get_mut::<PlayerState>(player_entity).unwrap();
@@ -324,31 +339,31 @@ pub fn update_player_inventory(world: &mut World) {
                 println!("WARNING: {}", err);
             }
         } else {
-            let mut draw_order = world.get_mut::<DrawOrder>(entity).unwrap();
-            draw_order.0 = ITEMS_DRAW_ORDER;
+            let mut drawable = world.get_mut::<Drawable>(entity).unwrap();
+            drawable.draw_order = ITEMS_DRAW_ORDER;
 
             let mut body = world.get_mut::<PhysicsBody>(entity).unwrap();
 
             body.is_deactivated = false;
 
-            let mut sprite_set = world.get_mut::<AnimatedSpriteSet>(entity).unwrap();
+            if let DrawableKind::AnimatedSpriteSet(sprite_set) = drawable.kind.borrow_mut() {
+                sprite_set.restart_all();
 
-            sprite_set.restart_all();
+                let sprite = sprite_set.map.get_mut(SPRITE_ANIMATED_SPRITE_ID).unwrap();
 
-            let sprite = sprite_set.map.get_mut(SPRITE_ANIMATED_SPRITE_ID).unwrap();
+                if sprite
+                    .animations
+                    .iter()
+                    .any(|a| a.id == *GROUND_ANIMATION_ID)
+                {
+                    sprite.set_animation(GROUND_ANIMATION_ID, true);
+                } else {
+                    sprite.set_animation(IDLE_ANIMATION_ID, true);
+                }
 
-            if sprite
-                .animations
-                .iter()
-                .any(|a| a.id == *GROUND_ANIMATION_ID)
-            {
-                sprite.set_animation(GROUND_ANIMATION_ID, true);
-            } else {
-                sprite.set_animation(IDLE_ANIMATION_ID, true);
-            }
-
-            if let Some(sprite) = sprite_set.map.get_mut(EFFECT_ANIMATED_SPRITE_ID) {
-                sprite.is_deactivated = true;
+                if let Some(sprite) = sprite_set.map.get_mut(EFFECT_ANIMATED_SPRITE_ID) {
+                    sprite.is_deactivated = true;
+                }
             }
         }
     }
